refactor(libs): update libs exports & move utility functions
re #72
