+++ apps/api/src/app/invites/invites.controller.ts
@@ -8,7 +8,7 @@ import {
     UseGuards
 } from "@nestjs/common"
 import { AuthGuard } from "@nestjs/passport"
-import { mapEntity } from "../common/utils"
+import { mapEntity } from "@zk-groups/utils"
 import { CreateInviteDto } from "./dto/create-invite.dto"
 import { Invite } from "./entities/invite.entity"
 import { InvitesService } from "./invites.service"
