+++ libs/onchain/src/index.ts
@@ -1,2 +1,2 @@
-export * from "./lib/ZKGroups/updateOffchainGroups"
-export * from "./lib/Semaphore/onchainGroups"
+export * from "./ZKGroups/updateOffchainGroups"
+export * from "./Semaphore/onchainGroups"
