+++ libs/utils/src/request.ts
@@ -1,7 +1,7 @@
 import axios, { AxiosRequestConfig, AxiosResponse } from "axios"
 
 /* istanbul ignore next */
-export async function request(
+export default async function request(
     url: string,
     config?: AxiosRequestConfig
 ): Promise<any> {
